refactor(lssutil): move utility functions into lssutil
